A sqrt that doesn't set errno
A standards-compliant implementation of sqrt() will bounds-check it
and set errno for invalid values - which is a significant overhead
when a fast implementation of sqrt available.

One way to avoid the overhead is to compile with a flag like
-fno-math-errno. Another is to use a custom sqrt() function, which is
what this change does.

This change reduces runtime by ~5% for acos and asin, tested with the
range of values that would not early out.